# COMPAT: Argsort position matches NumPy