--- a/pandas/core/indexes/datetimelike.py
+++ b/pandas/core/indexes/datetimelike.py
@@ -192,7 +192,7 @@ class DatetimeIndexOpsMixin(ExtensionIndex):
             return sorted_index, _as
         else:
             values = self._data
-            sorted_values = values[values.argsort()].asi8
+            sorted_values = np.sort(values.view("M8[ns]")).view("i8")
 
             freq = self.freq
             if freq is not None and not is_period_dtype(self):