--- a/pandas/tests/indexes/datetimes/test_ops.py
+++ b/pandas/tests/indexes/datetimes/test_ops.py
@@ -247,7 +247,7 @@ class TestDatetimeIndexOps(Ops):
             ),
             (
                 [pd.NaT, "2011-01-03", "2011-01-05", "2011-01-02", pd.NaT],
-                [pd.NaT, pd.NaT, "2011-01-02", "2011-01-03", "2011-01-05"],
+                ["2011-01-02", "2011-01-03", "2011-01-05", pd.NaT, pd.NaT],
             ),
         ],
     )
@@ -269,14 +269,17 @@ class TestDatetimeIndexOps(Ops):
         ordered, indexer = index.sort_values(return_indexer=True)
         tm.assert_index_equal(ordered, expected)
 
-        exp = np.array([0, 4, 3, 1, 2])
+        if index.isna().any():
+            exp = np.array([3, 1, 2, 0, 4])
+        else:
+            exp = np.array([0, 4, 3, 1, 2])
         tm.assert_numpy_array_equal(indexer, exp, check_dtype=False)
         assert ordered.freq is None
 
         ordered, indexer = index.sort_values(return_indexer=True, ascending=False)
         tm.assert_index_equal(ordered, expected[::-1])
 
-        exp = np.array([2, 1, 3, 4, 0])
+        exp = exp[::-1]
         tm.assert_numpy_array_equal(indexer, exp, check_dtype=False)
         assert ordered.freq is None