@TomAugspurger 2020-01-22
Perf looks OK I think

```python
import pandas as pd
import numpy as np

arr = np.arange(10000)
np.random.shuffle(arr)
idx = pd.DatetimeIndex(arr)

%timeit idx.sort_values()

idx = pd.PeriodIndex(ordinal=arr, freq='D')
%timeit idx.sort_values()
```

Master

```
403 µs ± 10.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
450 µs ± 9.41 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

PR

```
421 µs ± 3.39 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
468 µs ± 8.29 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

Not sure why they aren't identical yet. We're a bit outside the standard deviation on those runs.