@jbrockmendel 2020-01-23
> Not sure why they aren't identical yet. We're a bit outside the standard deviation on those runs.

i think we get one, maybe two extra function-call-overheads; does that account for 15-20 microseconds?