@TomAugspurger 2020-01-23
> im a bit confused by this. how is this prevented from varying based on the numpy version?

That was my [original question](https://github.com/pandas-dev/pandas/pull/31210#issue-365909033). Do we want to match the installed version of NumPy so that `np.sort(ndarray[m8[ns]])` matches `np.sort(DatetimeIndex), regardless of the NumPy versoin`? Or  do we want to adopt the new behavior of sorting last regardless of the NumPy version?

Both seem quite difficult, given that we sometimes do things like sort the datetimes, extract the i8 values, and then operate on them assuming they're sorted. I'm not confident that I can get this done for 1.0, if ever.