@jbrockmendel 2020-01-24
> I'm not confident that I can get this done for 1.0, if ever.

This does sound like the kind of problem that goes away on its own once our numpy min-version catches up.  I lean towards "dont sweat it"