@TomAugspurger 2020-01-27
I'm going to close this to get it out of the way.