@TomAugspurger 2020-01-27
Anything that relied on `.sort_values().asi8` to be sorted was broken / incorrect. This includes `.rolling`, `.resample`, etc.