@jbrockmendel 2020-01-28
Also affected: the ability to have DatetimeIndexOpsMixin.argsort dispatch to self._data.argsort